Filler uses normal constructor, fixed loop bug
Now the filler doesn't use reflection to get the constructor, and I fixed a bug with the filler loop. It was using x - 256 to check if the Item was null: Which meant for slot 390 (Which isn't used in vanilla minecraft), it would check to see if slot 134 was empty, and since that slot wouldn't be empty (There's a block there) it create a filler block at 390 when it shouldn't. The bugfix was, of course, removing the - 256, which is only to be used when you want to create an Item at a specific slot.