Patches for firing events in previous commit.
Altered:

World.spawnEntityInWorld() - added a simple type check to throw
ItemSpawnEvent when EntityItems are passed in.  Cancelling the event
stops the entity being added to the world.

EntityItem - added an additional field, lifespan, which is defaulted to
6000.  This field is used when checking for item expiration, rather than
being hardcoded to 6000.

EntityItem.onUpdate() - expiration comparison now uses lifespan field.
When expiration is triggered, an ItemExpireEvent is fired before the
EntityItem is flagged as dead.  Cancelling the event stops the call to
setDead()

EntityPlayer.dropPlayerItemWithRandomChoice() - added firing for both
ItemTossEvent and ItemDropEvent.  When the random motion flag is true,
the items are being dropped on death.  When false, the items are being
thrown out.  Cancelling either event prevents the EntityItem from
spawning in the World.

ItemSeeds.tryPlaceIntoWorld() - tangentially related.  I wanted to test
events by having sseds plant themselves on soil when they expire.
Simplfied the method a bit, and added ability to pass in null as the
PlayerEntity.