PlayerDropsEvent replaces ItemDropEvent
Replaced ItemDropEvent with a child class of LivingDropsEvent which
seemed like a better solution to the problem.  PlayerDropsEvent is fired
a single time when a player dies and includes all EntityItems about to
enter the world as well as additional information regarding how the
player died.

As creating this event required passing some extra data around
(DamageSource) through the item dropping code, I went ahead and cleaned
most of it up and added javadocs.