Implemented restrictive inventory mechanism.
The new interface IRestrictiveInventory provides a way to ask an
inventory whether it allows a specified item stack in a given slot. As
the Slot class already has such a query, which returns tre by default, I
patched it to use the interface if available. I also implemented the
interface on TileEntityFurnace and TileEntityBrewingStand to match the
gui behaviour.

If properly respected by modders, this will solve the infinite xp issue
https://github.com/MinecraftForge/MinecraftForge/issues/161