Update ShapedOreRecipe to fix constructors.
This removes the initial constructors from ShapedOreRecipe
to fix problems with constructor ambiguity. The detecting
of if it should be miriored is handled in the main
constructor, and still defaults to true. This should fix issue
MinecraftForge/MinecraftForge#208

Adds Javadoc, and changes mirriored to mirrored.

Fixes checkMatch variable names and setMirriored.

setMirriored  is depracated and replaced with setMirrored.