Mods which extend Slot can use custom textures for slot background overlay, instead of only items.png
This could be implemented more generally to allow all Slot to have a custom texture and index specified for background instead of requiring a subclass of it to override the getters, but it's my first time submitting anything to Forge and I didn't want it to be too complicated and risk breaking others' code. The mod I wanted this change for already extends Slot with several subclasses, so this is good enough.

If the more general implementation is desired, contact me - I can try writing something - or others are welcome to improve upon this submission.

Signed-off-by: Yancarlo Ramsey <yancarlo@gmail.com>