Slot BG texture/index assignment; inventory index from private to protected
Expanded Slot class improvements to allow any slot to have a custom background texture and index assigned without requiring the use of a subclass to override getters. Also changed the one private field to protected so that subclasses don't have to store a duplicate of it through their constructors. (All other fields were public, including the IIventory reference... but not the index.)