Forge Additions: Conditional Recipe
This patch add the support for Conditional Recipes.
Conditional recipes can be used with a Forge Hook or Event.
The Hook is useful for "weekdays modders", which want something limited
but functional: you can use a default set of conditions like real and
minecraft world time, AOI Logical operators, location, player and
entity.
The Event is useful if you need "something more".