Vanilla bugfix for entity collision above/below world
Entities do not collide when they're above Y=255, or below Y=0.
To test this, fly at Y=300 or Y=-30, and try hitting yourself in the
head with a bow by shooting upwards gently.
This is due to the for loop in Chunk.getEntitiesWithinAABBForEntity
not executing its body.

The related method Chunk.getEntitiesOfTypeWithinAAAB does not seem to
need this fix.