Fix calculation of movement speed
The distance used to calculate the movement speed check should be the
maximum moved, not the minimum.
This makes sense due to the fact that to figure out if someone is moving
too fast, you should see how fast they moved at their fastest point, and
not at their slowest point.
This fixes the bug where a player can teleport without permissions with
a modified client without getting caught for moving too fast.