Added the OreStack class
The OreStack class is for mod recipes that support stack sizes larger
than 1, and I think that it would be convenient to have that in forge.
It allows sizes different than 1 for OreDictionary names in recipes for
things like the Ic2 Compressor or the RP2 Alloy furnace

p.s.: I'm bad at explanations.