Hook for rendering item as bow and event for player facing enderman
The first hook allows modders to let one of their items (preferably a
custom bow) render /being held as a bow, without a need to make a custom
Item Renderer.
The second is an event which is triggered whenever a player is facing /
looking at an Enderman. It is useful as an event because modders are
able to check the circumstances for if the Enderman should get
aggressive towards the looking player, without needing to have an item /
block for it.