Initial commit for 1.6.1
To make sure we are compliant with MCP policy, I have moved to a patch
system based on MinecraftForge. Our old repository has been renamed to
MCPC-Plus-Legacy due to the amount of changes for 1.6. All MCPC changes to
MC source will now be found in the patches folder.

For developers:
In order to setup the new MCPC+ eclipse environment, run setup.py to
automatically configure and build it. After the environment is setup, if
you want to modify existing MC sources, use the Forge/Minecraft sources
for any changes then run update_patches.py to generate the new set of
patches used for your PR. All non-mc sources can be changed directly and
PR'd as is.