Simple way to control the bottom position of the chat window.
There is no simple way to change the vertical position of the chat
window, which is needed if a custome health bar is used that has a
different size, like in TerraFirmaCraft mod. The chat window is being
rendered behind the health bar covering a part of the last sentence.
This small change would make things a lot easier.