Disable CB's item attributes patch. Fixes #26, #82, #84, #104, #106
With 1.6.2, CB added a patch to store item attributes in their
CraftItemMeta class. This worked fine for vanilla items but obviously
didn't support modded items. It will be temporarily disabled until I can
provide a proper fix. Any item that was used prior to this will have to be
remade since NBT was stripped by this patch.