Prevent subsequent listeners from acting on a canceled event
This change prevents a situation where one listener cancels an event and a subsequent listener is still able to act on the event (by failing to check the isCanceled() value).  An example of this could be if a mod is using the BreakEvent event of one block as a catalyst for taking other actions such as destroying an entire tree. My understanding is that only the original block break would be canceled and any other actions taken would still be allowed. Adding the check inside the loop prevents that sort of asshattery.