Send a Packet103SetSlot to client when a BlockPlaceEvent is cancelled.
Currently, whenever a player places a block in a protected area the
equipped itemstack size on client is never updated properly since the client
*thinks* the block was placed. The reason this happens is because
ItemStack.areItemStacksEqual returns true since the server does not
decrement stack size if a BlockPlaceEvent is cancelled. This causes
NetServerHandler.handlePlace not to send the appropriate packet to client which causes
the bug. To fix this, I simply force a Packet103SetSlot to be sent
whenever a block placement fails.