Fixed TechneModel Rendering
The TechneModel was previously broken in the way it rendered the models. I patched it up by fixing up the rotations and translations of the model as well as added in a setup() method (Line 262) to invert the model by default.

With this model loader, the player can just go:

model.renderAll()

And the model will come out all smooth. Already tested with multiple mods in 1.6.4 and 1.7.2.