Update forge_at.cfg
Currently, entities only drop experience when killed by a player. setting the getExperiencePoints method to public would allow for modders to get the experience when the mob is killed by a custom entity. This would also prevent the modder from needing to use playerdamage on the mob, which would cause it to attack the player, and not the entity itself. (example : levelable pets)