Added ArmorChangedEvent
Added an event to fire upon changes to the armor slots. It is posted to
the event bus when onSlotChanged is called on the four armor slots in
ContainerPlayer.
By this, developers will be able to recognize these changes more easily,
without comparing armor on each tick, and possibly force players to drop armor or call other effects upon equipping.