Change method of deciding if a Block is eligible to be a Beacon's base block
This adds a new List that is generated on startup, and can later be added to or have blocks removed from it by a mod.

This will come in handy for, say, (nameless mod that adds Diamond Bricks :)), allowing these to be used as a beacon base instead of just fizzling out.

I made no other changes except for two lines, yet git diff shows some weird output. I'm pushing this onto GitHub to see if it makes more sense there

Signed-off-by: Cruz Julian Bishop <cruz@techern.com>