Re-Add Branch Logic to Shearable Blocks on harvest
This is to fix issue #1308

This re-adds the logic to detect if a block is being harvested by a
shear tool. Removed the code that used to run so now if I shear a Leaf
block, the Ishearable logic is executed and the onblockharvested logic
is skipped.

Re-Add Branch Logic to Shearable Blocks

now with added support for custom shears