Added PlayerChangeCurrentItemEvent
This event is called when the player change his current item. It can be
useful when  the player is using an item and then change his current
item (the event PlayerUseItemEvent.Stop isn't called).

I placed the event's post in "if" blocks so if you want to cancel this
event in the future, it will be easier.

The slot index represents the future slot index.