VillagerRegistry changes
Adds the ability to add custom villager professions and careers, as well
as the ability to add trades for certain levels of the careers.

the random villager selector will now randomly select any profession
added, and the random career selector will properly look for the careers
available to that profession from the registry.

The sample mod works a follows:
right click a villager with a stick prints out his profession and career
number.

Shift right click with a stick spawns in a villager with the newly added
profession, though using spawn eggs will also occasionally get one.