Villagers now properly handles adding and removing professions
If a profession is removed, any villager with that profession has levels
and trades reset, and is reset to farmer.
If a profession's id is changed, it automatically re-assigns it
accordingly, so professions can be added or removed as needed, with
trades only changing if the villager's profession is removed.

Vanilla spawning works as it used to, with any building/spawnegg that
used to make random villagers anyway being able to have the new ones as
well, but any building with specific spawns, still only makes that
specific spawn, this is ensured so long as the new villagers are
registered after the vanilla ones.

The test mod adds three new professions, and it was tested by randomly
commenting some of them out, then running minecraft, then commenting
some of them back in.  The villagers which were not removed properly
kept their professions, and the ones that were removed were set to
farmers.