Changes VillagerCareer to allow creation of custom careers.
Changes when VillagerCareer.trades is initialized, so that the private
init() doesn't need to be called.
Also updates the test mod to showcase the creation of a new career, the
programmer. The name is not displayed correctly in the gui. This is
because of the way Vanilla determines the display name. This would be a
task for another pull request.