Generalizes villager names and prevents collisions
This uses the VillagerCareer objects to determine the villagers
unlocalized name and allows one to add careers to vanilla professions
and getting the correct names to be displayed.

Also, collisions are prevented by adding the domain name of the mod
owning the profession to the unlocalized name.

The test mod is updated to test the features mentioned above.