Changes unlocalized name format for careers (again...)
Now the unlocalized name follows the format:
entity.Villager.<domain>.<profession_name>.<career_name>

Naming conflicts should now be much rarer and names for villager more
easily understandable, as the unloc. name contains the mod's id. Also,
the career is an extension to the profession, so therefore it should be
included in the villager name.