Intended to solve conflicts between mods that add inventories for the player and make a tab for them in the survival inventory. also allows for less key-bind usage.
added survival tabs that can be registered instantiating them.
registers their individual id depending on the container they have been attached too.
Only works for GuiContainers.

added SurvivalTablist class, an encasing class that handles all the drawing, button, and page logic to keep the code in GuiContainer minimal.

patched GuiInventory to call the super for actionPerformed, so it will pick up the button's from the survival tabs

patched guicontainer with a few simple lines that refer to the SurvivalTabsList class, a class that handles all logic and drawing for the survival tabs.

gitignore