Rewrites sync(), TypeAdapters and adds map test to test mod.
The code for syncing maps got massive, but it now allows users to add
new keys to the config files, which then will be available to the mod.

Also, the new annotations @RequiresMcRestart and @RequiresWorldRestart
allow the modder to specify, that change to a property a category
requires the specified restart.