Tweak Event.setCanceled To Match Convention
It is generalized Java convention to throw a UnsupportedOperationException whenever an operation is not supported, as in the case of Event.setCanceled() being called on an event where Event.isCancelable() returns false. The prior implementation threw an IllegalArgumentException, which did not make sense, given that it was not the argument that was the issue, but the fact that the method was even called (given that the method, as per its current implementation ignores the value of its parameter when determining if the exception should be thrown). Also, in the hopes of improving logs, the exception message has been changed as to include the canonical name of the Event subclass instance that was not cancelable. Thus, these changes should aid modders in determining the exact path (or mod interaction) that caused the given event instance as to be spawned and then subsequently attempted as to be canceled.