Improve Exception Message Wording
This commit clarifies the wording of the message associated with the UnsupportedOperationException thrown by the Event#setCanceled() method when the given event instance is non-cancelable. Furthermore, this commit also corrects the missing second star from the JavaDoc for the aforementioned method that had turned the JavaDoc into a block comment instead. Thank you @mezz for your assistance with this patch.