Create event for overriding mob griefing behaviour
Create an event which is fired when the mobGriefing game rule's value is
checked. The event will allow the game rule to be overridden by any
event listeners handling the event.

Create EntityMobGriefingEvent, extending EntityEvent so that listeners
can make decision based on entity information.

Update ForgeEventFactory to add a getMobGriefingEvent method, this will
replace each instance of retrieving the mobGriefing game rule value.
The result of the event will be handled here to either override the
behaviour or use the default.

Create EntityMobGriefingEventTest to allow easy testing of the event
handling. This will allow the event result to be set using name tags.

Create and update the patch files to replace all instances of the
retrieval of the mobGriefing game rule value with the new event.