Added documentation, fixes inconsistencies and related bug.
The interfaces IFieldWrapper and ITypeAdapter were poorly documented, which was changed in this commit.
Especially IFieldWrapper was problematic, as the fully qualified property name was given as parameter "name" or "key" (-> inconsistency). Also, sometimes these parameters ment the suffix or the fully qualified name. This was fixed.
The related parts in ConfigManager were adjusted to these changes.

ConfigManager::sync was further documented, so that other programmers will have an easier time to debug and fix this code.