Fix for the player inadvertently being removed from the Entity Tracker when the chunk they were in unloads after they teleport out of it.
This changes the chunk unload to send a forced update to any players that are listed as being in the chunk. If the player is no longer in the chunk this will cause them to update their location and properly register with the chunk they are in.

This is needed because in 1.10 the order things were done during a tick was changed. Prior to this, entities got an update prior to any chunks unloading, which gave them a chance to "move out" of the chunk and into a new one. As of 1.10, chunk unloads happen prior to updating entities, so any entity still listed as being in an unloading chunk will be removed from the entity tracker.

This causes a variety of issues including players being unable to get out of beds, and invisible players.

It is possible this needs to be extended to all entities rather than just players, but player entities are the ones with the most noticeable issues due to the change in tick order.