Added WatcherPotionEffect and LastHealth
WatcherPotionEffect is a watcher added to make sure that the client and
server are always in agreement with which potion effects are active. The
client does not influence what potion effects the server has on an given
entity.

LastHealth acts similarly to Heath as it is always one step behind what
health was. This data is persistent and written to the nbt.