Added WatcherPotionEffect and LastHealth
WatcherPotionEffect is a watcher added to make sure that the client and
server are always in agreement with which potion effects are active. The
client does not influence what potion effects the server has on any
given
LivingEntity.

LastHealth acts similarly to Heath as it is always one step behind what
health was. This data is persistent and written to the nbt and a getter
and setter were added.

I have created mod that uses both of these features called Damage
Indicator that uses these values to draw the entities health above its
head as it would look for the player, including potion effects.

LastHealth is needed so that a ghostly heart can be drawn where the
health was at before the previous injury.

I have tested as well as I am able and have not found any problems.

Here is a forum post as to why I felt that these additions were
necessary as well as some pictures of the mod. Note: I have changed some
of the code given in that post to better integrate it into the forge
code and required fewer changes than I originally thought.