Remove FluidStack amount from hashcode calculation
The current implementation of FluidStack#hashcode and FluidStack#equals breaks the implementation contract for these methods.

Alternatively, this could be solved by directing FluidStack#equals to the FluidStack#isFluidStackIdentical function. Though this is probably more likely to break existing code than the proposed solution of removing the amount from the hashcode calc.

I'm not specifically aware of what effects this bug has, I only discovered it via code inspection rather that debugging. So I'm not particularly fussed about the outcome of this PR, but I felt it at least deserved a discussion.