Move registry events to after preinit
Registry events should remain in the same place they were in 1.12.2 to avoid confusion.  While there is no benefit to either position, modders expect to be able to use @EventBusSubscriber and/or register handlers for registry events to the bus in preinit.  The confusion that would be created as a side effect of leaving the registry events before preinit is unnecessary.