fml|ModInfo.java: `logoFile` wrong order of priority
the "logoFile" of the `mods.toml` template in MDK does not work.  It
seems that there are 3 cases from where to take it's value. The case
`!logoFile.isPeresent()` was the second and
`modConfig.getOptional("logoFile")` had the lowest priority (the case
of MDK template).

Now `modConfig.getOptional("logoFile")` has highest priority and
overrides all. `!logoFile.isPeresent()` has lowest priority now.