Fix error prone state lookup during rail neighbor updates.
In a delayed neighbor notification via pistons, the requested block
state to perform the neighbor update can already be "outdated" and no
longer be the same block type at the given position. Vanilla does not
encounter this issue because it simply performs the rail direction
lookup on the IBlockState being passed in, whereas here, the requested
block state of the current position from the live World may already be
different from this block instance. This was initially reported on
SpongeForge:
https://github.com/SpongePowered/SpongeForge/issues/2920
and later proven to be a Forge specific bug:
https://www.minecraftforge.net/forum/topic/74448-crash-with-piston-and-rails/

The workaround here is to get the current up-to-date IBlockState at the
position, without directly passing it in to the getRailDirection added
method, checking that the current state's Block instance is the same as
this, and if so, passing the current state. A small optimization to
avoid duplicate state lookups for the same position is also made to find
out if the block itself is air. Again, because vanilla does not perform
the forge-added block state lookup, vanilla's logic is already only
performing a single lookup. Here, we can save the extra cost of a
duplicative lookup by storing the current state in a local variable.

The benefit of the change is that the original logic is still in tact,
due to the air block sanity check after the flag has been resolved to
true or not.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>