Added isBubbleSource and getDrag in IForgeBlock instead of using an interface and patching vanilla Soul Sand and Magma Block.
This way the classes SoulSandBlock and MagmaBlock won't get patched and the BubbleColumnBlock patch is as close to the vanilla version as possible.

I've tested it with a block, which extends SoulSandBlock and overrides isBubbleSource to return true and getDrag to return false.
I didn't put the test into the pull request since I don't think this needs further testing.