Clearer method names and descriptions.
Title says it all. Plus I've used the Direction Property, which already
was imported in both IForgeBlock and BubbleColumnBlock to make it
clearer in which direction the resulting bubble elevator will push you.

My opinion is, that any modder, who wants a more complicated behavior
of the bubble column block, can add a modded version and create that
one with his 'bubble source block' instead of the vanilla one.