GH-6330 Don't log query commands
CommandSource#sendFeedback(ITextComponent, bool) will send feedback to
the console and logged-in ops if the bool flag is set. Looking at
Minecraft, it seems Mojang draw the line between commands that query
state and those that change it - this patch makes an equivalent change
to Forge.

Specifically the following commands no longer log feedback:
- /forge dimensions
- /forge entity list
- /forge mods
- /forge tps