Allows json based biomes to be loaded.
Previously they would error because only the biomes loaded from the Biome Registry would have the forge:registry_name property injected into the biome json.
Error loading registry data: No key forge:registry_name in MapLike[...]
This is because there are 2 implementations of WorldSettingsImport.IResourceAccess, RegistryAccess which was used when you open the new world screen and loaded from the registry, and an anonymous class in WorldSettingsImport.IResourceAccess which was used for datapacks.
RegistryAccess had been patched to inject forge:registry_name property into the json.
This patches the anonymous class to do the same.

Signed-off-by: Justin Wiblin <justin_wiblin@hotmail.co.uk>