Only apply the MC-68754 fix on Windows, closes #7546 (#7549)
MC-68754 only affects Minecraft on Windows, as per the comments on Mojang's bug tracker, and the LWJGL bug report (https://github.com/LWJGL/lwjgl/issues/142). This change simply checks if the OS running Minecraft is Windows before applying the fix. This was tested on Windows 10, MacOS Mojave, and Lubuntu (5.8.0-31), and all platforms allowed the window to be resized as expected after exiting fullscreen. As the MC-68754 fix was the cause of #7546, this issue is fixed by applying this patch.