[1.16.x] Add IForgeItem#getMobToSpawn
The spawn egg item has been problematic to users for a number of reasons, specifically its desire to rely on an IdentityHashMap making implementations of supplied entity types even more complex than it should be. Technically the spawn egg could be represented as a new Item besides for one method, `#getChildToSpawn`. This method is called within `MobEntity` to spawn a child when a mob is right clicked with its associated spawn egg. This could be handled with `PlayerInteractEvent$EntityInteract`; however, the optional supplied calls a protected method that would need to be accessed using reflection for AT.

To get around this to make spawn eggs fully implementable from the base item, a new method is added. This essentially takes in a new mob entity to create and spawns it in the world. From there, the optional is mapped to its corresponding entry and calls the protected method. Finally, if its present, it returns SUCCESS on the server. Note that this method is called only on the logical server similarly to the `SpawnEggItem` implementation.

As an alternative solution, I could instead make the method an interaction for all mob entities that returns an `ActionTypeResult` circumventing the need for the event altogether besides for vanilla items interacting with vanilla entities. This still makes it impossible for the protected method to be called without AT. However, the only method call this would be applicable to is within `FoxEntity`. It more or less depends on how much should the entity interaction with items be reliant on the event.